Refactor: use return value to indicate a shutdown, avoid using ServerState::Shutdown.
Try not to use ServerState as workflow control.
Using it introduces state checking everywhere.
Just return an Err when state changes.
